refactor(canon): split sfc_typecheck into cohesive modules
Extract the ~1.4k-line inline #[cfg(test)] mod tests block out of
sfc_typecheck.rs into a dedicated sfc_typecheck/tests.rs submodule,
leaving the entry file as a thin (~53-line) module declaration that
re-exports the existing analysis/runner items unchanged.

The four insta snapshots are moved alongside the relocated test source
(into sfc_typecheck/snapshots/) so insta resolves them from the test
file's directory; module path (sfc_typecheck::tests) and snapshot names
are unchanged. Pure structural move: no logic, signature, or public API
changes.
